+++ database/migrations/2022_06_05_100053_create_arbitro_disciplinas_table.php
@@ -15,8 +15,8 @@ class CreateArbitroDisciplinasTable extends Migration
     {
         Schema::create('arbitros_disciplinas', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('arbitro_id');
-            $table->unsignedInteger('campeonato_id');
+            $table->unsignedBigInteger('arbitro_id');
+            $table->unsignedBigInteger('campeonato_id');
             $table->foreign('arbitro_id')->references('id')->on('arbitros');
             $table->foreign('campeonato_id')->references('id')->on('campeonatos');
             $table->timestamps();
